+++ tests/conftest.py
@@ -4,12 +4,31 @@ from collections import namedtuple
 
 import botocore.session
 from botocore.stub import Stubber
+import pytest
 from pytest import fixture
 
 
 def pytest_addoption(parser):
     parser.addoption('--skip-slow', action='store_true',
                      help='Skip slow tests')
+
+
+def pytest_configure(config):
+    config.addinivalue_line("markers", "slow: mark test as slow to run")
+    config.addinivalue_line(
+        "markers", (
+            "on_redeploy: mark an integration test to be run after "
+            "the app is redeployed"
+        )
+    )
+
+
+def pytest_collection_modifyitems(config, items):
+    if config.getoption("--skip-slow"):
+        skip_slow = pytest.mark.skip(reason="Skipping due to --skip-slow")
+        for item in items:
+            if "slow" in item.keywords:
+                item.add_marker(skip_slow)
 
 
 @fixture(autouse=True)